frontend: Cleanup frontend code stuff.
This PR cleans up some technical debt around the layout of the
frontend code.

Includes:
- Adds a proper directory structure.
- Support for linting, formatting, tools etc.
- Support for test coverage

Required for:
- https://github.com/gerrymandr/precinct-explorer/issues/9

/cc @radiantradon
