+++ frontend/.prettierrc
@@ -1,0 +1,5 @@
+printWidth: 100
+semi: true
+singleQuote: true
+trailingComma: none
+bracketSpacing: true
