+++ frontend/Makefile
@@ -1,0 +1,15 @@
+.PHONY: run
+run:
+	npm start
+
+.PHONY: install
+install:
+	npm install
+
+.PHONY: format
+format:
+	npm run format
+
+.PHONY: lint
+lint:
+	npm run lint
