+++ frontend/package.json
@@ -1,22 +1,42 @@
 {
+  "name": "precinct-explorer",
+  "description": "Frontend for Precinct Explorer",
+  "license": "MIT",
+  "version": "0.0.1",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/gerrymandr/precinct-explorer.git"
+  },
   "scripts": {
+    "format": "prettier --write \"{src,examples,test}/**/*.js\" --loglevel warn",
+    "lint": "prettier-check \"{src,test}/**/*.js\" && eslint src test",
     "start-local": "webpack-dev-server --env.local --progress --hot --open",
-    "start": "webpack-dev-server --progress --hot --open",
-    "format": "./node_modules/prettier/bin-prettier.js app.js --write"
+    "start": "webpack-dev-server --progress --hot --open"
   },
   "dependencies": {
     "d3-request": "^1.0.5",
-    "deck.gl": "^5.0.0",
+    "deck.gl": "^5.1.4",
     "deck.gl-layers": "0.0.24",
-    "luma.gl": "^5.1.3",
+    "luma.gl": "^5.1.4",
+    "npm": "^5.8.0",
     "prettier": "^1.11.1",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
     "react-dropdown": "^1.4.0",
-    "react-map-gl": "^3.2.0"
+    "react-map-gl": "^3.2.5",
+    "react-popper": "^0.8.3",
+    "react-transition-group": "^2.2.1",
+    "reactstrap": "^5.0.0-beta.3"
   },
   "devDependencies": {
     "buble-loader": "^0.4.0",
+    "eslint": "^4.13.1",
+    "eslint-config-prettier": "^2.9.0",
+    "eslint-config-uber-es2015": "^3.1.2",
+    "eslint-config-uber-jsx": "^3.3.3",
+    "eslint-plugin-react": "^7.5.1",
+    "prettier": "^1.9.2",
+    "prettier-check": "^2.0.0",
     "webpack": "^2.2.0",
     "webpack-dev-server": "^2.2.0"
   }
