+++ frontend/webpack.config.js
@@ -7,7 +7,7 @@ const webpack = require("webpack");
 
 const CONFIG = {
   entry: {
-    app: resolve("./app.js")
+    app: resolve("./src/app.js")
   },
 
   devtool: "source-map",
@@ -39,5 +39,4 @@ const CONFIG = {
 };
 
 // This line enables bundling against src in this repo rather than installed deck.gl module
-module.exports = env =>
-  env ? require("../webpack.config.local")(CONFIG)(env) : CONFIG;
+module.exports = env => env ? require("../webpack.config.local")(CONFIG)(env) : CONFIG;
